+++ backend/src/main/java/org/example/coretrack/model/material/inventory/MaterialInventory.java
@@ -1,5 +1,6 @@
 package org.example.coretrack.model.material.inventory;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -19,9 +20,9 @@ public class MaterialInventory {
     @JoinColumn(nullable = false, name = "materialVariant_id")
     private MaterialVariant materialVariant;
 
-    private Integer currentStock;
-    private Integer minAlertStock;
-    private Integer maxStockLevel;
+    private BigDecimal currentStock;
+    private BigDecimal minAlertStock;
+    private BigDecimal maxStockLevel;
 
     @OneToMany(mappedBy = "materialInventory", cascade = CascadeType.ALL)
     List<MaterialInventoryLog> logs = new ArrayList<>();
@@ -44,7 +45,7 @@ public class MaterialInventory {
     public MaterialInventory(){
     }
 
-    public MaterialInventory(MaterialVariant materialVariant, Integer currentStock, Integer minAlertStock, Integer maxStockLevel,
+    public MaterialInventory(MaterialVariant materialVariant, BigDecimal currentStock, BigDecimal minAlertStock, BigDecimal maxStockLevel,
             List<MaterialInventoryLog> logs, User created_by) {
         this.materialVariant = materialVariant;
         this.currentStock = currentStock;
@@ -74,27 +75,27 @@ public class MaterialInventory {
         this.materialVariant = materialVariant;
     }
 
-    public Integer getCurrentStock() {
+    public BigDecimal getCurrentStock() {
         return currentStock;
     }
 
-    public void setCurrentStock(Integer currentStock) {
+    public void setCurrentStock(BigDecimal currentStock) {
         this.currentStock = currentStock;
     }
 
-    public Integer getMinAlertStock() {
+    public BigDecimal getMinAlertStock() {
         return minAlertStock;
     }
 
-    public void setMinAlertStock(Integer minAlertStock) {
+    public void setMinAlertStock(BigDecimal minAlertStock) {
         this.minAlertStock = minAlertStock;
     }
 
-    public Integer getMaxStockLevel() {
+    public BigDecimal getMaxStockLevel() {
         return maxStockLevel;
     }
 
-    public void setMaxStockLevel(Integer maxStockLevel) {
+    public void setMaxStockLevel(BigDecimal maxStockLevel) {
         this.maxStockLevel = maxStockLevel;
     }
